Worked a bit more on the Plan D
I added the assets to the sync, including the script and resource files
(buymenu.res, weapon_plan_d.txt and dab_english.txt) I had to copy over
the pistol animations for the third person model to animate - we need to
add the PlanD to the activities list of the player model I think.

I haven't done any gameplay programming, it is just implemented as a new
weapon. I imagine this gun being whipped out when G is pressed and you
should be able to fire immediately using LMB, then when it is empty you
drop it, no reloading. You can switch back to your main weapon with a
bullet still in the spout for use later. I have it set to 48 damage,
because I can't figure out how to give it damage drop off. I also need
to prevent it from auto-reloading when empty.